requirements.txt
smolagents[openai]>=1.24.0
openai>=1.30.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.5.0
matplotlib>=3.8.0
Pillow>=10.0.0
PyYAML>=6.0.0
ultralytics>=8.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
datasets>=2.18.0
huggingface_hub>=0.24.0
kaggle>=1.6.0
selenium>=4.15.0
tqdm>=4.66.0
python-dotenv>=1.0.0
imagehash>=4.3.1
nbformat>=5.10.0
catboost>=1.2.0
pytest>=8.0.0
